pyproject.toml
[project]
name = "health-dispatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.192.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.2",
    "python-jobspy>=1.1.82",
]
